+++ app/Http/Controllers/Admin/PlaceController.php
@@ -2,17 +2,21 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Helpers\DeleteImage;
+use App\Helpers\ImageResize;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Admin\City;
 use App\Models\Admin\Place;
 use App\Models\Admin\Activity;
+use Alert;
+use App\Models\Admin\Event;
 
 class PlaceController extends Controller
 {
     public function index(Place $place)
     {
-        $places = Place::all();
+        $places = Place::orderBy('created_at', 'desc')->get();
         return view('admin.place.index', compact('places'));
     }
 
@@ -34,10 +38,12 @@ class PlaceController extends Controller
         $place->visitation = $request->place['visitation'];
         $place->description = $request->place['description'];
         $place->image = $request->place['image']->store('places');
+
+        ImageResize::reduce($place->image);
         $place->save();
+        $place->activities()->sync((array) $request->input('place.activity'));
 
-        $place->activities()->sync((array)$request->input('place.activity'));
-
+        Alert::success(trans('adminlte::pages.messages.saved'));
         return redirect(route('admin.places.show', compact('place')));
     }
 
@@ -51,7 +57,15 @@ class PlaceController extends Controller
 
     public function destroy(Place $place)
     {
-        $place->delete();
+        if (Event::where('place_id', $place->id)->count()) {
+            Alert::info(trans('adminlte::pages.messages.not_allowed'));
+            return back();
+        } else {
+            DeleteImage::unlink($place->image);
+            $place->delete();
+            Alert::success(trans('adminlte::pages.messages.deleted'));
+        }
+
         return redirect(route('admin.places.index'));
     }
 
@@ -59,16 +73,24 @@ class PlaceController extends Controller
     {
         $this->validation($request);
 
+        if (isset($request->place['image'])) {
+            DeleteImage::unlink($place->image);
+
+            $place->place_image = $request->place['image']->store('places');
+            ImageResize::reduce($place->image);
+        }
+
         $place->name = $request->place['name'];
         $place->city_id = $request->place['city'];
         $place->location = $request->place['location'];
         $place->visitation = $request->place['visitation'];
         $place->description = $request->place['description'];
-        $place->place_image = isset($request->place['image']) ? $request->place['image']->store('places') : null;
+
+        ImageResize::reduce($place->image);
         $place->update();
+        $place->activities()->sync((array) $request->input('place.activity'));
 
-        $place->activities()->sync((array)$request->input('place.activity'));
-
+        Alert::success(trans('adminlte::pages.messages.updated'));
         return redirect(route('admin.places.show', compact('place')));
     }
